requirements.txt
hbmqtt
python-dotenv
